+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### ✅ Added
 - Added factory method for customizing the message list container's modifier
+- Option to customize the date separation logic in the message list
 
 # [4.33.0](https://github.com/GetStream/stream-chat-swiftui/releases/tag/4.33.0)
 _June 09, 2023_